// src/schemas/interfaces/services.d.ts
export interface IServiceSelectableFields {
  name?: number;

  vendor?: number;

  state?: number;

  description?: number;

  category?: number;

  country?: number;

  loc?: number;

  address?: number;

  city?: number;

  date?: number;

  capacity?: number;

  suspended?: number;

  photos?: number;

  price?: number;

  isActive?: number;

  isDeleted?: number;

  deletedAt?: number;

  bannerImages?: number;
}